+++ resources/lang/en/validation/deleteAccount.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'password.required' => 'Please fill in your current password.',
+    'password.current_user_password' => 'The password is incorrect.'
+];