+++ resources/lang/fr/validation/deleteAccount.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'password.required' => 'Veuillez saisir votre mot de passe actuel.',
+    'password.current_user_password' => 'Le mot de passe est incorrect.'
+];